Modularized Config class
