+++ .gitignore
@@ -1,4 +1,4 @@
-config.json
+config.txt
 .vscode
 build
 accounts.txt
