+++ snipe.py
@@ -1,3 +1,5 @@
+from config import Config
+
 try:
     import aiohttp
     import logging
@@ -24,16 +26,6 @@ logging.basicConfig(level=logging.INFO, format='%(message)s')
 times = []
 global sent_reqs
 sent_reqs = 0
-default_config = """timing_system:api
-skin:https://namemc.com/skin/4e45a59795a252c3
-skin_model:slim
-change_skin:false
-snipe_reqs:8
-block_reqs:3
-auth_delay:800
-max_accs:30
-auto_link_namemc:false NOT IMPLEMENTED
-"""
 
 
 def custom_info(message):
@@ -237,57 +229,6 @@ async def time_snipe(target, block_snipe):
                 except Exception:
                     print(f"{Fore.WHITE}[{Fore.RED}{Fore.RED}]{Fore.RESET} Failed to time snipe")
     return timing
-
-
-class Config:
-    def __init__(self):
-        self.options = []
-        if not os.path.exists("config.txt"):
-            with open("config.txt", "w") as f:
-                f.write(default_config)
-        with open("config.txt", "r") as f:
-            unconverted_lines = f.readlines()
-            self.lines = list()
-            for line in unconverted_lines:
-                self.lines.append(line.strip())
-            self.timing = self.find_parameter("timing_system")
-            self.block_reqs = int(self.find_parameter("block_reqs"))
-            self.snipe_reqs = int(self.find_parameter("snipe_reqs"))
-            self.skin = self.find_parameter("skin")
-            self.max_accs = int(self.find_parameter("max_accs"))
-            if "namemc.com/skin" in self.skin:
-                self.skin = f"https://namemc.com/texture/{self.skin.split('/')[-1]}.png"
-            self.custom_announce = self.find_parameter("custom_announce")
-            if self.custom_announce is None:
-                del self.custom_announce
-            self.change_skin = self.find_bool("change_skin", False)
-            self.auth_delay = int(self.find_parameter("auth_delay"))
-            self.webhooks = self.find_all("wh")
-            self.skin_model = self.find_parameter("skin_model")
-
-    def find_parameter(self, parameter):
-        for line in self.lines:
-            line = line.split(":")
-            if line[0].lower() == parameter:
-                line.pop(0)
-                line = ":".join(line)
-                self.options.append(line)
-                return line
-
-    def find_bool(self, parameter, default):
-        parameter = self.find_parameter(parameter)
-        parameter = {"false": False, "true": True}[parameter]
-        return parameter
-
-    def find_all(self, parameter):
-        options = []
-        for line in self.lines:
-            line = line.split(":")
-            if line[0] == parameter:
-                line.pop(0)
-                options.append(":".join(line))
-                self.options.append(options)
-        return options
 
 
 class Account:
